feat: move multi-tab chat app to ChatApp hidden behind beta flag and restore AgentApp
