+++ src/Ivy.Tendril/AppShell/TendrilAppShell.cs
@@ -103,7 +103,8 @@ public class TendrilAppShell(AppShellSettings settings) : ViewBase
             ["icebox"] = status.IceboxCount,
             ["recommendations"] = status.RecommendationsCount,
             ["trash"] = status.TrashCount,
-            ["agent"] = status.ActiveAgentChatsCount
+            ["agent"] = status.ActiveAgentChatsCount,
+            ["chat"] = status.ActiveAgentChatsCount
         };
         var agentId = config.Settings.CodingAgent;
         return repo.GetMenuItems()
