+++ src/Ivy.Tendril/BetaAppAttribute.cs
@@ -1,0 +1,4 @@
+namespace Ivy.Tendril;
+
+[AttributeUsage(AttributeTargets.Class)]
+public class BetaAppAttribute : Attribute;
