+++ src/Ivy.Tendril/Constants.cs
@@ -16,6 +16,7 @@ public static class Constants
     public const int PullRequests = 60;
     public const int Icebox = 70;
     public const int Agent = 80;
+    public const int Chat = 85;
     public const int Trash = 90;
     public const int Help = 100;
     public const int Onboarding = 110;
